Furniture System Changes
To add furniture to the editor window, use the MetaInformation object.

BasicEditor is replaced by FurnitureEditorScript.

Furniture now has type IDs, and there is a
Furniture.InstantiateFurnitureByID () method.

There is a Game script that will contain all savefile-related
information.

There is a MetaInformation script that contains all information
pertinent to the game project (for example, it maps furniture IDs to
prefabs, and keeps track of the Room prefab).